# docs/public/api/openapi.yaml
openapi: 3.1.0
info:
  title: CometCMS Public API
  version: "1.0.0"
  description: Stable HTTP API for headless content, media, and integration scripts. Workspace-scoped routes are available under /api/v1/workspaces/{workspace} with the same content, schema, and media paths.
servers:
  - url: https://yourdomain.com/api/v1
security: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API token created in the admin UI with permission grants.
  parameters:
...
    Filename:
      name: filename
      in: path
      required: true
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
      description: Maximum number of items. Omit to return all matching items.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
      description: Number of matching items to skip.
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
    ListMeta:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        sort:
          type: string
        order:
          type: string
          enum: [asc, desc]
...
    MediaItem:
      type: object
      required:
        [filename, name, url, size, mime, category, alt, title, visibility]
      properties:
        filename:
          type: string
        name:
          type: string
        url:
          type: string
          format: uri
        size:
          type: integer
          minimum: 0
        mime:
          type: string
        category:
          type: string
        alt:
          type: string
        title:
          type: string
        visibility:
          type: string
          enum: [public, private]
        width:
          type: [integer, "null"]
          minimum: 1
        height:
          type: [integer, "null"]
          minimum: 1
        uploaded_by:
          type: [string, "null"]
        uploaded_at:
          type: [string, "null"]
          format: date-time
    Ok:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
paths:
...
  /media:
    get:
      tags: [Media]
      summary: List media files
      description: Without a token, only public files are returned. With `media.read` on `media:*`, private files are included.
      x-required-permission:
        effect: allow
        actions: [media.read]
        resources: ["media:*"]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Media file list.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MediaItem"
                  meta:
                    allOf:
                      - $ref: "#/components/schemas/ListMeta"
                      - type: object
                        properties:
                          categories:
                            type: array
                        items:
                          type: string
        "401":
          description: Invalid bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Token missing `media.read` permission.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [Media]
      summary: Upload media
      description: Requires `media.upload` on `media:*`, or `media:category:{category}` when assigning a category.
      x-required-permission:
        effect: allow
        actions: [media.upload]
        resources: ["media:*", "media:category:{category}"]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [media]
              properties:
                media:
                  type: array
                  items:
                    type: string
                    format: binary
                category:
                  type: string
      responses:
        "201":
          description: Uploaded media items.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MediaItem"
                  meta:
                    type: object
                    properties:
                      categories:
                        type: array
                        items:
                          type: string
        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Token missing `media.upload` permission.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Upload or validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /media/{filename}:
    delete:
      tags: [Media]
      summary: Delete a media file
      description: Requires `media.delete` on `media:{filename}`.
      x-required-permission:
        effect: allow
        actions: [media.delete]
        resources: ["media:{filename}"]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Filename"
      responses:
        "200":
          description: File deleted.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Ok"
        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Token missing `media.delete` permission for the file.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"